+++ web/src/components/integrations/todoist/notification_list_item.rs
@@ -1,8 +1,6 @@
 #![allow(non_snake_case)]
 
-use chrono::{DateTime, Local};
 use dioxus::prelude::*;
-
 use dioxus_free_icons::{icons::bs_icons::BsCardChecklist, Icon};
 
 use universal_inbox::{
@@ -11,11 +9,14 @@ use universal_inbox::{
     HasHtmlUrl,
 };
 
-use crate::components::{
-    integrations::todoist::{icons::Todoist, list_item::TodoistListItemSubtitle},
-    list::{ListContext, ListItem},
-    notifications_list::{get_notification_list_item_action_buttons, TaskHint},
-    Tag, TagDisplay,
+use crate::{
+    components::{
+        integrations::todoist::{icons::Todoist, list_item::TodoistListItemSubtitle},
+        list::{ListContext, ListItem},
+        notifications_list::{get_notification_list_item_action_buttons, TaskHint},
+        Tag, TagDisplay,
+    },
+    utils::format_elapsed_time,
 };
 
 #[component]
@@ -25,11 +26,7 @@ pub fn TodoistNotificationListItem(
     is_selected: ReadOnlySignal<bool>,
     on_select: EventHandler<()>,
 ) -> Element {
-    let notification_updated_at = use_memo(move || {
-        Into::<DateTime<Local>>::into(notification().updated_at)
-            .format("%Y-%m-%d %H:%M")
-            .to_string()
-    });
+    let notification_updated_at = use_memo(move || format_elapsed_time(notification().updated_at));
     let list_context = use_context::<Memo<ListContext>>();
     let task_icon_style = match todoist_item().priority {
         TodoistItemPriority::P1 => "",